Tùy biến và deploy mô hình/docker_sample/server/requirements.txt
torch==1.12.0
python-dotenv==0.21.0
Pillow==7.1.2
numpy
flask-cors==3.0.10
flask
torchvision==0.13.0